LPS-137270 Extract component to its own file, no logic changes
