+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/_CollectionFilterConfigurationModal.scss
@@ -2,8 +2,8 @@
 
 html#{$cadmin-selector} {
 	.cadmin {
-		.page-editor__collection-general-panel {
-			&__modal .modal-body {
+		.page-editor__collection-filter-configuration-modal {
+			& .modal-body {
 				padding: 0;
 			}
