+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/_Plugin.scss
@@ -1,8 +1,8 @@
-@import 'item-configuration-panels/CollectionGeneralPanel';
 @import 'item-configuration-panels/FieldSet';
 @import 'item-configuration-panels/MappingPanel';
 @import 'item-configuration-panels/RowStylesPanel';
 
+@import 'CollectionFilterConfigurationModal';
 @import 'ItemConfiguration';
 @import 'PageStructureSidebar';
 @import 'PageStructureSidebarSection';
